Inventory tab: environment selector
Middleware / BMSAPI:
Add hasInventory field to StudyInstance. Refactor query so that the same
join can be used for the field and for the canBeDeleted field. Re-format
query for readability.
Leave a note to improve the hasFieldLayout/hasGeoJSON fields (not a
problem now because it works for the plot dataset)

Fieldbook:
Get study instances and show the ones that hasInventory

Issue: IBP-3752