If available, use OAuth token, else X-Auth-Token
OAuth token format is only limited to brapi now, but there is one
scenario where the url is not /brapi: spring error handling where a
second request to /error passes through the filter, and in that case the
oauth token should be used too, if available, as is the case for brapi
endpoints. If not, the request will not pass authentication and the
response will be the one of Http401UnauthorizedEntryPoint.
This problem doesn't manifest in swagger because it sends both headers.

To overcome this, we switch to a workflow where the oauth token is used
if the header is available, and if not we fallback to x-auth-token.

This shouldn't be a problem because internally both tokens are the same
and the idea is to move to oauth for bmsapi eventually.

Issue: IBP-3901