Germplasm search > add to list > select all
Now that table is limited to max 5000 entries, it's possible to add a
"select all" checkbox for this finite list.

Use the select all to send searchRequest to save list.

Workbench:
- Implement select all
- Add generics to search-composite.ts for type checking

BMSAPI:
- PagedResult: remove default assignment filteredResults <- totalResults
when filtered is 0. It makes some sense when filtered is not
implemented, but when it is implemented, it doesn't hold because
filtered could actually be zero.
- GermplamListServiceImpl: get entries from composite

Middleware:
- Add composite to GermplasmListGeneratorDTO
- Allow null pageable in dao method to get all max results (5000)

Issue: IBP-4114