IBP-5790: Remove Pedigree level=40 in COP, retrieve full pedigree
If pedigrees exceed this level, or if the full pedigree is not
retrieved, the `order` property of the nodes will be incorrect. If it is
 consistently incorrect among all nodes (all order relative to the same
 ancestor) , calculation may be ok anyway,
 but if it's not, this could cause that the cop expansion is not being
 done in the right order (https://github.com/IntegratedBreedingPlatform/Middleware/blob/97756bca6609c930782abc744e0e8e98132954b0/src/main/java/org/generationcp/middleware/api/germplasm/pedigree/cop/CopCalculation.java#L176)
The max level was initially set because of some test data with very
large pedigrees that could cause some OutOfMemoryError.
TODO a way to have both memory protection and accurate order of nodes.